+++ manager/it/util/src/main/java/org/opendaylight/vtn/manager/it/util/action/SetIcmp4CodeVerifier.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2015 NEC Corporation.  All rights reserved.
+ * Copyright (c) 2015 NEC Corporation. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v1.0 which accompanies this distribution,
@@ -23,7 +23,7 @@ public abstract class SetIcmp4CodeVerifier
      * @param code  The ICMP code to be set.
      */
     public SetIcmp4CodeVerifier(short code) {
-        super(Icmp4Factory.class, code);
+        super(Icmp4Factory.class, (int)code);
     }
 
     // ActionVerifier
@@ -35,7 +35,7 @@ public abstract class SetIcmp4CodeVerifier
     public void apply(EthernetFactory fc) {
         Icmp4Factory ic4fc = getFactory(fc, Icmp4Factory.class);
         if (ic4fc != null) {
-            ic4fc.setCode((byte)getValue());
+            ic4fc.setCode((short)getValue());
         }
     }
 }
