+++ manager/it/util/src/main/java/org/opendaylight/vtn/manager/it/util/action/SetTcpDstVerifier.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2015 NEC Corporation.  All rights reserved.
+ * Copyright (c) 2015 NEC Corporation. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v1.0 which accompanies this distribution,
@@ -21,7 +21,7 @@ public abstract class SetTcpDstVerifier extends SetTpDstVerifier<TcpFactory> {
      *
      * @param port  The destination port number to be set.
      */
-    public SetTcpDstVerifier(short port) {
+    public SetTcpDstVerifier(int port) {
         super(TcpFactory.class, port);
     }
