+++ manager/it/util/src/main/java/org/opendaylight/vtn/manager/it/util/action/SetTcpSrcVerifier.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2015 NEC Corporation.  All rights reserved.
+ * Copyright (c) 2015 NEC Corporation. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v1.0 which accompanies this distribution,
@@ -21,7 +21,7 @@ public abstract class SetTcpSrcVerifier extends SetTpSrcVerifier<TcpFactory> {
      *
      * @param port  The source port number to be set.
      */
-    public SetTcpSrcVerifier(short port) {
+    public SetTcpSrcVerifier(int port) {
         super(TcpFactory.class, port);
     }
